Dev monitoring statistics and time series (#222)
* separated default monitoring implementation in its own file

* started implementation of statistics

* started generating JSON output for statistics from monitor

* improved test for monitoring

* removed margo_set_monitor

* added name and config in monitor struct and changed the way configuration is managed

* changed margo_set/get_monitoring_data to attach data to handles instead of requests

* lots of statistics added for clients

* added statistics collection for origin and target

* added statistics collection for bulk operations

* added functions to convert monitoring structures to json

* added bulk statistics in JSON and updated unit test

* added full generation of JSON output from statistics

* completed generation of JSON from monitoring and unit test

* added parent ID tracking using ABT key

* added unit test for parent RPC id

* updated to more recent version of uthash

* started adding code to track sender and receiver addresses

* clang-formatting uthash.h so clang-format stops complaining

* implemented per-source/destination RPC statistics

* added create_size attribute for bulk statistics

* done with per-address statistics generation

* fixed unit test for margo monitoring

* cleaned up monitoring code and added per-hash mutex

* added session pools to remove allocation from critical paths

* added timeout value statistics for progress

* added missing mutex locks when writing the JSON file

* started moving ABT profiling into its own system

* removed old diagnostics code, and cleaned up a lot of the code

* removed the need for extra space in input buffer

* added time series for RPC counts

* added time series for bulk sizes

* added time series for pool size and pool total size

* added unit test for time series

* added more comments

* removed unnecessary configuration code in margo-init.c

* trying to make codeql happy

* more changes for codeql

* fixed bug in progress loop sampling

* added MARGO_ENABLE_MONITORING environment variable and fixed problem with ref count of handle in on_rpc_ult

* cleaner way of handling handle refcount in RPC ULT

* fixed pretty_json property in unit test

* only triggers with actual_count > 0 are now counted by statistics

* removed unnecessary comment

* added self address in monitoring files

* fixed memory leaks

* added step in github action to run with ASAN

* changed the handler statistics to also have timestamp (though unused)

* reorganized the JSON format for bulk create operations

* reorganized the JSON format for bulk transfer operations

* some renaming of JSON sections and some nesting

* added some comments in margo-monitoring.h

* removed margo-gen-profile

* removed margo-gen-profile from scripts that need to be installed

* replace parent_id with parent_rpc_id

* removed old doc

* changed relative_time_from_handler and relative_time_from_create for some of the operations

* changed relative timestamps to get more precise timings

* actually take into account enable_statistics and enable_time_series (not just when writing files)

* adding msec to progress_timeout_value

* changed some keys in JSON

+++ src/margo-macros.h
@@ -153,7 +153,7 @@ inline static struct json_object* json_object_copy(struct json_object* in)
             if (!json_object_is_type(_tmp, json_type_int))                     \
                 MARGO_WARNING(0, "Overriding field \"%s\" with value %d",      \
                               __field_name, (int)__value);                     \
-            else if (json_object_get_int64(_tmp) != __value)                   \
+            else if (json_object_get_int64(_tmp) != (int64_t)(__value))        \
                 MARGO_WARNING(0, "Overriding field \"%s\" (%d) with value %d", \
                               __field_name, json_object_get_int64(_tmp),       \
                               __value);                                        \
